Move themeBtnHandler() into theme.js
