+++ assets/javascript/theme.js
@@ -40,4 +40,20 @@ const lightTheme = (btnEl) => {
   }
 }
 
-export { lightTheme, darkTheme }
+function themeBtnHandler() {
+  // If no theme in localStorage, use timeOfDay() to decide theme
+  const currentTheme = JSON.parse(localStorage.getItem('currentTheme')) || timeOfDay();
+  const nextTheme = 'LIGHT' === currentTheme ? 'DARK' : 'LIGHT';
+
+  // the user is saving the date object to a variable
+  const date = new Date();
+  // then pulling the hours from the date object
+  const hours = date.getHours();
+
+  // toggle theme of the webpage
+  nextTheme === "DARK"
+    ? darkTheme(this)
+    : lightTheme(this)
+}
+
+export { lightTheme, darkTheme, themeBtnHandler }
